requirements.txt
Django
psycopg2-binary
python-decouple
djangorestframework
dj-rest-auth
django-filter
gunicorn